+++ CHANGELOG.md
@@ -1,3 +1,55 @@
+# Version 0.3.5
+
+## Bugfixes
+
+* Fixed poblem with plain integer values within param values
+
+## Features
+
+* Usage of several backreferences per param value are now possible
+* Added MPEventConversion connector which allows for the tracking of unique users within event hit types
+
+# Version 0.3.4
+
+## Bugfixes
+
+* Wrong usage of CURLs POST fields
+
+## Features
+
+* None
+
+# Version 0.3.3
+
+## Bugfixes
+
+* None
+
+## Features
+
+* Added different hit types and extended error handling
+* Added GA cookie inclusion for the Client ID
+
+# Version 0.3.2
+
+## Bugfixes
+
+* None
+
+## Features
+
+* Allow for client IP forwarding within the analytics module measurement protocol connector
+
+# Version 0.3.1
+
+## Bugfixes
+
+* Updated dependencies for the webserver.json example configuration
+
+## Features
+
+* Added an analysis module
+
 # Version 0.3.0
 
 ## Bugfixes